Add function that connects mongo, gets collection and converts id to mongo object id
